reviews: Revert rename
