+++ bin/faucet/src/errors.rs
